JFA Voroni Filter
The Voroni filter needs an input set of points in image form, the point
color needs to work with the function (r=x, g=y, b=quadrant).  It will
generate a diagram of colors that represent the center point for their
quadrant.  The consumer filter takes that and uses it with an image to
do something very similar to the crystallize filter of Core Image (OS X
only)