Added kGPUImageRotate180.
This is needed for the back facing camera on landscape right, since the back camera is left facing landscape.

NOTE: Since a 180 rotation is a flip vert and a flip horz, and both of those are marked as broken in GPUImageCropFilter's -calculateCropTextureCoordinates, so is the additional block to handle rotate180.