prepareToCaptureImage race condition with createFilterFBOofSize addressed
Rather than set the important preparedToCaptureImage ivar, it sets a prepareToCaptureImage ivar that createFilterFBOofSize then uses to understand that it should at that point prepare and then set preparedToCaptureImage.

Note that prepareToCaptureImage also destroys the outputTexture. My understanding is that outputTexture shouldn't be set unless the image pipeline has been started and in which case the code is potentially vulnerable to a race condition. This isn't necessarily true if the pipeline is stopped… but perhaps this is an okay restriction? I've added a warning NSLog at the moment. As such this is more of a suggestion of the implementation rather than perhaps a permanent solution in this part.