If GPUImagePicture is passed an empty image, create a small 2x2 image to prevent a possibly fatal error on CGContextDrawImage
The empty image passed to GPUImagePicture will trigger this error when CGContextDrawImage is called on init::

"<Error>: CGContextDrawImage: invalid context 0x0. This is a serious error. This application, or a library it uses, is using an invalid context  and is thereby contributing to an overall degradation of system stability and reliability. This notice is a courtesy: please fix this problem. It will become a fatal error in an upcoming update."

This should prevent the error altogether.