GPUImageMovieWriter: fix bug when writing a movie with audio
The audio was not written with dispatch_sync so the sample buffer was invalidated before it was written, which caused the AVAssetWriter to error.

Also improved the error logging if there is an AVAssetWriter error to include the error itself, as well as detecting audio writing errors and logging them.