GPUImageMovieWriter: fix bug when writing a movie with audio
Change from previous fix. We do want to allow audio to be written asynchronously, so instead we no longer invalidate the audio buffer in GPUImageMovie. Invalidating the buffer does not appear to be necessary.

Invalidating the buffer causes writing audio to fail because the buffer is invalidated by the time we come to write it to the AVAssetWriter asynchronously.