Turn off front-camera workaround on iOS 6 or later
As discussed in #462, there are various problems with the workaround for corrupt images on iOS 5. Given that iOS 5 is quite legacy now, I think it's a reasonable solution to turn off that workaround if GPUImage is running on iOS 6 or later.

This could of course be adjusted if the precise version of the OS required for the workaround is not 5.X.