Lots going on here, but there was no logical way to separate it into smaller changes.
Allowing GPUImageMovie to take an AVPlayerItem for playback as well as an AVAsset in order to allow it to play AVComposition objects. When sourcing from an AVPlayerItem, it now uses CADisplayLink to handle the timing to ensure proper A/V sync.

Adding GPIImageMovieComposition specifically for reading from an AVCompostion object in non-real time. This is primarily so it can be connected to a GPUImageMovieWriter in order to render a composition to file.

Lots of changes to how audio/videoInputCallback(s) are managed to make rendering to a file reliable.