GPUImageMovieWriter: synchronise finishRecording with the movieWritingQueue
As the AVAssetWriter gets started on the movieWritingQueue.

Also ensure that the handler is called async, otherwise it is called on the movieWritingQueue, which is dispatch_sync from the video processing queue, so if the user-code in the handler calls anything that dispatch_syncs to the video processing queue: deadlock.