[IMAGE-54] Intermediate state where reconstruction for one image is bit exact
When you select the HDR Debug mode and don't touch the slider or click,
the left part of the screen is the original image, the right part of the screen is
the laplacian reconstruction with the original weight map.

For one image, it has to be bit-identical, which it is.

There are still memory management issues with the various debug modes.